refactor(migration-sdk-viem): rename file
