+++ packages/migration-sdk-viem/src/positions/supply/aaveV3Optimizer.supply.ts
@@ -28,7 +28,10 @@ import {
 import { signTypedData } from "viem/actions";
 import { morphoAaveV3Abi } from "../../abis/aaveV3Optimizer.js";
 import { getMorphoAaveV3ManagerApprovalTypedData } from "../helpers/signatures.js";
-import { MigratableSupplyPosition } from "./index.js";
+import {
+  type IMigratableSupplyPosition,
+  MigratableSupplyPosition,
+} from "./index.js";
 
 interface IMigratableSupplyPosition_AaveV3Optimizer
   extends Omit<IMigratableSupplyPosition, "protocol"> {
